Creation of test driver
Devices and drivers are used to communicate with the external world and therefore produce non-deterministic events. This makes testing the virtual device interface harder. This commit creates a determinist test driver that is used for testing purposes only.
