+++ firmware/device/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "device"
+version = "0.1.0"
+edition = "2021"
+
+[[bin]]
+name = "device"
+path = "main.rs"
+
+[dependencies]
+miralis_abi = { path = "../../crates/abi" }
+log = { workspace = true }
