+++ justfile
@@ -45,6 +45,7 @@ test:
 	cargo run --package runner -- run --config {{qemu_virt}} --firmware os_ecall
 	cargo run --package runner -- run --config {{qemu_virt}} --firmware vectored_mtvec
 	cargo run --package runner -- run --config {{qemu_virt_release}} --firmware default
+	cargo run --package runner -- run --config {{qemu_virt_release}} --firmware device
 
 	# Testing with Miralis as firmware
 	cargo run --package runner -- run --config {{qemu_virt}} --firmware miralis
