+++ src/platform/visionfive2.rs
@@ -11,7 +11,9 @@ use crate::arch::{Arch, Architecture};
 use crate::config::{
     PLATFORM_NB_HARTS, TARGET_FIRMWARE_ADDRESS, TARGET_STACK_SIZE, TARGET_START_ADDRESS,
 };
-use crate::device::{self, VirtClint};
+use crate::device::clint::{VirtClint, CLINT_SIZE};
+use crate::device::tester::{VirtTestDevice, TEST_DEVICE_SIZE};
+use crate::device::{self, VirtDevice};
 use crate::driver::ClintDriver;
 use crate::{Platform, _stack_start, _start_address};
 // —————————————————————————— Platform Parameters ——————————————————————————— //
@@ -21,6 +23,7 @@ const MIRALIS_START_ADDR: usize = TARGET_START_ADDRESS;
 const FIRMWARE_START_ADDR: usize = TARGET_FIRMWARE_ADDRESS;
 
 const CLINT_BASE: usize = 0x2000000;
+const TEST_DEVICE_BASE: usize = 0x3000000;
 
 // ———————————————————————————— Platform Devices ———————————————————————————— //
 
@@ -32,6 +35,8 @@ static CLINT_MUTEX: Mutex<ClintDriver> = unsafe { Mutex::new(ClintDriver::new(CL
 
 /// The virtual CLINT device.
 static VIRT_CLINT: VirtClint = VirtClint::new(&CLINT_MUTEX);
+/// The virtual test device.
+static VIRT_TEST_DEVICE: VirtTestDevice = VirtTestDevice::new();
 pub static WRITER: Mutex<Writer> = Mutex::new(Writer::new(SERIAL_PORT_BASE_ADDRESS));
 
 // ———————————————————————————————— Platform ———————————————————————————————— //
@@ -96,13 +101,22 @@ impl Platform for VisionFive2Platform {
         usize::MAX
     }
 
-    fn create_clint_device() -> device::VirtDevice {
-        device::VirtDevice {
+    fn create_virtual_devices() -> [VirtDevice; 2] {
+        let virtual_clint: device::VirtDevice = VirtDevice {
             start_addr: CLINT_BASE,
-            size: device::CLINT_SIZE,
+            size: CLINT_SIZE,
             name: "CLINT",
             device_interface: &VIRT_CLINT,
-        }
+        };
+
+        let virtual_test_device: device::VirtDevice = VirtDevice {
+            start_addr: TEST_DEVICE_BASE,
+            size: TEST_DEVICE_SIZE,
+            name: "TEST",
+            device_interface: &VIRT_TEST_DEVICE,
+        };
+
+        [virtual_clint, virtual_test_device]
     }
 
     fn get_clint() -> &'static Mutex<ClintDriver> {
